requirements.txt
olefile==0.44
pandoc-latex-environment==0.0.4
pandocfilters==1.4.1
Pillow==4.0.0
pip==9.0.1
pypandoc ==1.3.3
PyQt5==5.7.1
setuptools==32.3.1
sip==4.19
wheel==0.29.0